move toast logic to hook
