+++ src/components/Toast.tsx
@@ -1,225 +1,30 @@
-import React, {
-  useState,
-  useEffect,
-  useRef,
-  isValidElement,
-  cloneElement
-} from 'react';
+import React, { isValidElement, cloneElement } from 'react';
 import cx from 'classnames';
 
 import { ProgressBar } from './ProgressBar';
 import { WithInjectedOptions } from '../types';
-import { NOOP, canUseDom, RT_NAMESPACE, isFn } from '../utils';
+import { NOOP, RT_NAMESPACE, isFn } from '../utils';
 
 import { TransitionProps } from 'react-transition-group/Transition';
-import { useToast, useKeeper } from '../hooks';
-
-type DragEvent = MouseEvent & TouchEvent;
-
-function getX(e: DragEvent) {
-  return e.targetTouches && e.targetTouches.length >= 1
-    ? e.targetTouches[0].clientX
-    : e.clientX;
-}
-
-function getY(e: DragEvent) {
-  return e.targetTouches && e.targetTouches.length >= 1
-    ? e.targetTouches[0].clientY
-    : e.clientY;
-}
-
-const iLoveInternetExplorer =
-  canUseDom && /(msie|trident)/i.test(navigator.userAgent);
-
-interface Draggable {
-  start: number;
-  x: number;
-  y: number;
-  deltaX: number;
-  removalDistance: number;
-  canCloseOnClick: boolean;
-  canDrag: boolean;
-  boundingRect: DOMRect | null;
-}
+import { useToast } from '../hooks';
 
 export const Toast: React.FC<WithInjectedOptions> = props => {
-  const { isRunning, pauseToast, playToast } = useToast(props);
-  const [preventExitTransition, setPreventExitTransition] = useState(false);
-  const toastRef = useRef<HTMLDivElement>(null);
-  const prevProps = useKeeper({
-    draggable: props.draggable,
-    pauseOnFocusLoss: props.pauseOnFocusLoss
-  });
-  const drag = useKeeper<Draggable>({
-    start: 0,
-    x: 0,
-    y: 0,
-    deltaX: 0,
-    removalDistance: 0,
-    canCloseOnClick: true,
-    canDrag: false,
-    boundingRect: null
-  });
-
-  useEffect(() => {
-    if (isFn(props.onOpen))
-      props.onOpen(isValidElement(props.children) ? props.children.props : {});
-    if (props.draggable) bindDragEvents();
-    if (props.pauseOnFocusLoss) bindFocusEvents();
-
-    return () => {
-      if (isFn(props.onClose))
-        props.onClose(
-          isValidElement(props.children) ? props.children.props : {}
-        );
-      if (props.draggable) unbindDragEvents();
-      if (props.pauseOnFocusLoss) unbindFocusEvents();
-    };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  useEffect(() => {
-    if (prevProps.draggable !== props.draggable) {
-      props.draggable ? bindDragEvents() : unbindDragEvents();
-    }
-
-    if (prevProps.pauseOnFocusLoss !== props.pauseOnFocusLoss) {
-      props.pauseOnFocusLoss ? bindFocusEvents() : unbindFocusEvents();
-    }
-
-    prevProps.draggable = props.draggable;
-    prevProps.pauseOnFocusLoss = props.pauseOnFocusLoss;
-  }, [props]);
-
-  function bindFocusEvents() {
-    window.addEventListener('focus', playToast);
-    window.addEventListener('blur', pauseToast);
-  }
-
-  function unbindFocusEvents() {
-    window.removeEventListener('focus', playToast);
-    window.removeEventListener('blur', pauseToast);
-  }
-
-  function bindDragEvents() {
-    document.addEventListener('mousemove', onDragMove);
-    document.addEventListener('mouseup', onDragEnd);
-
-    document.addEventListener('touchmove', onDragMove);
-    document.addEventListener('touchend', onDragEnd);
-  }
-
-  function unbindDragEvents() {
-    document.removeEventListener('mousemove', onDragMove);
-    document.removeEventListener('mouseup', onDragEnd);
-
-    document.removeEventListener('touchmove', onDragMove);
-    document.removeEventListener('touchend', onDragEnd);
-  }
-
-  function onDragStart(
-    e: React.MouseEvent<HTMLElement, MouseEvent> | React.TouchEvent<HTMLElement>
-  ) {
-    const toast = toastRef.current!;
-    drag.canCloseOnClick = true;
-    drag.canDrag = true;
-    drag.boundingRect = toast.getBoundingClientRect();
-    toast.style.transition = '';
-    drag.start = drag.x = getX(e.nativeEvent as DragEvent);
-    drag.removalDistance = toast.offsetWidth * (props.draggablePercent / 100);
-  }
-
-  function onDragMove(e: MouseEvent | TouchEvent) {
-    const toast = toastRef.current!;
-
-    if (drag.canDrag) {
-      if (isRunning) pauseToast();
-
-      drag.x = getX(e as DragEvent);
-      drag.deltaX = drag.x - drag.start;
-      drag.y = getY(e as DragEvent);
-
-      // prevent false positif during a toast click
-      drag.start !== drag.x && (drag.canCloseOnClick = false);
-
-      toast.style.transform = `translateX(${drag.deltaX}px)`;
-      toast.style.opacity = `${1 -
-        Math.abs(drag.deltaX / drag.removalDistance)}`;
-    }
-  }
-
-  function onDragEnd() {
-    const toast = toastRef.current!;
-    if (drag.canDrag) {
-      drag.canDrag = false;
-
-      if (Math.abs(drag.deltaX) > drag.removalDistance) {
-        setPreventExitTransition(true);
-        props.closeToast();
-        return;
-      }
-
-      toast.style.transition = 'transform 0.2s, opacity 0.2s';
-      toast.style.transform = 'translateX(0)';
-      toast.style.opacity = '1';
-    }
-  }
-
-  function onDragTransitionEnd() {
-    if (drag.boundingRect) {
-      const { top, bottom, left, right } = drag.boundingRect;
-
-      if (
-        props.pauseOnHover &&
-        drag.x >= left &&
-        drag.x <= right &&
-        drag.y >= top &&
-        drag.y <= bottom
-      ) {
-        pauseToast();
-      } else {
-        playToast();
-      }
-    }
-  }
-
-  // Maybe let the end user tweak it later on
-  // hmmm no comment about ie. I hope this browser die one day
-  // don't want to fix the issue on this browser, my head is hurting too much
-  function onExitTransitionEnd() {
-    if (iLoveInternetExplorer) {
-      props.onExited!();
-      return;
-    }
-    const toast = toastRef.current!;
-    const height = toast.scrollHeight;
-    const style = toast.style;
-
-    requestAnimationFrame(() => {
-      style.minHeight = 'initial';
-      style.height = height + 'px';
-      style.transition = 'all 0.4s ';
-
-      requestAnimationFrame(() => {
-        style.height = '0';
-        style.padding = '0';
-        style.margin = '0';
-      });
-      setTimeout(() => props.onExited!(), 400);
-    });
-  }
-
+  const {
+    isRunning,
+    preventExitTransition,
+    toastRef,
+    onExitTransitionEnd,
+    eventHandlers
+  } = useToast(props);
   const {
     closeButton,
     children,
     autoClose,
-    pauseOnHover,
     onClick,
-    closeOnClick,
     type,
     hideProgressBar,
     closeToast,
-    transition: Foo,
+    transition,
     position,
     className,
     style,
@@ -233,33 +38,16 @@ export const Toast: React.FC<WithInjectedOptions> = props => {
     rtl,
     toastId
   } = props as Required<WithInjectedOptions>;
-
-  const toastProps: Record<string, string | Function> = {
-    className: cx(
-      `${RT_NAMESPACE}__toast`,
-      `${RT_NAMESPACE}__toast--${type}`,
-      {
-        [`${RT_NAMESPACE}__toast--rtl`]: rtl
-      },
-      className
-    )
-  };
-
-  if (autoClose && pauseOnHover) {
-    toastProps.onMouseEnter = pauseToast;
-    toastProps.onMouseLeave = playToast;
-  }
-
-  // prevent toast from closing when user drags the toast
-  if (closeOnClick) {
-    toastProps.onClick = (e: React.MouseEvent) => {
-      onClick && onClick(e);
-      drag.canCloseOnClick && closeToast();
-    };
-  }
-
+  const cssClasses = cx(
+    `${RT_NAMESPACE}__toast`,
+    `${RT_NAMESPACE}__toast--${type}`,
+    {
+      [`${RT_NAMESPACE}__toast--rtl`]: rtl
+    },
+    className
+  );
   const controlledProgress = !!progress;
-  const T = Foo as React.FC<Partial<TransitionProps>>;
+  const Transition = transition as React.FC<Partial<TransitionProps>>;
 
   function renderCloseButton(closeButton: any) {
     if (!closeButton) return null;
@@ -270,7 +58,7 @@ export const Toast: React.FC<WithInjectedOptions> = props => {
   }
 
   return (
-    <T
+    <Transition
       in={props.in}
       appear
       onExited={onExitTransitionEnd}
@@ -280,13 +68,10 @@ export const Toast: React.FC<WithInjectedOptions> = props => {
       <div
         id={toastId as string}
         onClick={onClick}
-        {...toastProps}
+        className={cssClasses}
+        {...eventHandlers}
         style={style}
         ref={toastRef}
-        onMouseDown={onDragStart}
-        onTouchStart={onDragStart}
-        onMouseUp={onDragTransitionEnd}
-        onTouchEnd={onDragTransitionEnd}
       >
         <div
           {...(props.in && { role: role })}
@@ -314,7 +99,7 @@ export const Toast: React.FC<WithInjectedOptions> = props => {
           />
         )}
       </div>
-    </T>
+    </Transition>
   );
 };
